% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.heavylasso}
\alias{coef.heavylasso}
\title{Extract Coefficients from a Heavy-Tailed Lasso Model}
\usage{
\method{coef}{heavylasso}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"heavylasso"}, typically returned by \code{\link{heavylasso}}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric vector of estimated regression coefficients.
}
\description{
This function extracts the estimated regression coefficients from a fitted
heavy-tailed lasso model.
}
\references{
Mai, T. T. (2025). Heavy Lasso: sparse penalized regression under heavy-tailed noise via data-augmented soft-thresholding. arXiv preprint arXiv:2506.07790.

Mai, T. T. (2025). Exponential Lasso: robust sparse penalization under heavy-tailed noise and outliers with exponential-type loss.
}
\author{
The Tien Mai (\href{mailto:the.tien.mai@fhi.no}{the.tien.mai@fhi.no})
}
